Removed dependency on XNAMaths
Found that my 'Common.h' had included a reference to DX11/XNAMaths
headers and made me dependent on XMFloats. I never really questioned
where they came from after I used them for the DX11 Renderer, which is
annoying.
Welp. Now I have my own set of 2D and 3D vectors lovingly called
'EnVector2' and 'EnVector3' respectively (The 'En' is for Engine!). At
the very least I get to add operators and member functions for getting
the magnitude/normalising them which is p. cool!

+++ Personal D3D Framework/Entity.h
@@ -15,7 +15,7 @@ public:
 	~Entity(void);
 	virtual void Update();
 	virtual bool OnEvent(IEvent* e);
-	void AddForce(EnVector3 direction, int power);
+	void AddForce(EnVector3 direction, float power);
 
 	UINT uID;
 	EnVector3 position;
